% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{add_sample_context}
\alias{add_sample_context}
\title{Add sample, treatment, and timepoint context to a table of measurements}
\usage{
add_sample_context(
  meas.table,
  dataset.name,
  include.subject.context = TRUE,
  include.response.data = TRUE,
  ...
)
}
\arguments{
\item{meas.table}{A \code{data.frame}, or a \code{matrix} which will be
converted to a \code{data.frame}, containing measurements,
as returned by functions like \code{get_all_measurements}}

\item{include.subject.context}{Whether to also include available information
about the subjects, beyond the subject.id}

\item{...}{Additional params, such as \code{database.name}, are passed to
\code{get_all_samples}}
}
\value{
Returns a \code{data.frame} containing the measurements in melted
  format along with the sample, timepoint, and treatment regimen contextual
  information.
}
\description{
Add sample, treatment, and timepoint context to a table of measurements
}
