% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{add_subject_context}
\alias{add_subject_context}
\title{Add subject context to a table of data}
\usage{
add_subject_context(
  meas.table,
  dataset.name,
  include.response.data = TRUE,
  ...
)
}
\arguments{
\item{meas.table}{A \code{data.frame}, or a \code{matrix} which will be
converted to a \code{data.frame}, containing measurements,
as returned by functions like \code{get_all_measurements}}

\item{...}{Additional params, such as \code{database.name}, are passed to
\code{get_all_samples}}
}
\description{
Add subject context to a table of data
}
