% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{add_tcr_context}
\alias{add_tcr_context}
\title{Add tcr context to a table of data with a tcr.id column}
\usage{
add_tcr_context(
  meas.table,
  dataset.name,
  assay.name,
  measurement.set.name,
  ...
)
}
\arguments{
\item{meas.table}{The table w/tcr.id that needs tcr data joined on to it.}

\item{dataset.name}{The name of the dataset which contains the tcrs}

\item{assay.name}{The name of the assay which contains the tcrs.}

\item{measurement.set.name}{The name of the measurement set which contains
the tcrs.}

\item{...}{Additional params, such as \code{database.name}, are passed to
\code{get_all_tcrs}}
}
\description{
Add tcr context to a table of data with a tcr.id column
}
