% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{add_variants_context}
\alias{add_variants_context}
\title{Add variants context}
\usage{
add_variants_context(meas.table, ...)
}
\arguments{
\item{meas.table}{A \code{data.frame} of measurements as returned by
\code{link{get_all_measurements}}}

\item{...}{Additional parameters passed to \code{\link{get_all_variants}}
e.g. \code{db.name}}
}
\value{
Returns the input data with additional information about the variants
}
\description{
This function adds additional variant information to a table of measurements
}
