% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{aggregate_otu_measurements}
\alias{aggregate_otu_measurements}
\title{Aggregate otu measurements at different taxonomic levels}
\usage{
aggregate_otu_measurements(
  tab,
  all.otus,
  cast.output = FALSE,
  normalize = TRUE,
  na.value = "Unclassified"
)
}
\arguments{
\item{tab}{OTU measurements as returned by \link{get_all_measurements}}

\item{all.otus}{A table with OTU information as returned by \link{get_all_otus}}

\item{normalize}{Whether OTU read counts should be normalized to the total read count
in each sample}

\item{na.value}{Name to use for OTU's that have an \code{NA} at a specific taxonomic level}

\item{cast.ouput}{If specified, the aggregated data is cast into a matrix, otherwise it will
be returned molten}
}
\value{
Returns a named list, with elements corresponding to the data aggregated
  at the corresponding taxonomic level. Note that in the aggregated data, the names
  of some taxa will appear as the concatenation of the names of upstream taxonomic levels.
  This is necessary because, in some cases, the name at each level is not enough to uniquely identify
  a taxonomic group (e.g. there could be multiple families with the same name in
  different orders)
}
\description{
This function aggregates OTU counts at different taxonomic level
to produce a table of per-taxa counts
}
