% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_adverse_events}
\alias{all_adverse_events}
\title{Query for all adverse events}
\usage{
all_adverse_events(dataset.name, clinical.observation.set.name = NULL)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{clinical.observation.set.name}{The name of the clinical-observation-set. All the AEs in that
set will be pulled}
}
\value{
Returns a query suitable for getting all AEs
  for a specified dataset
}
\description{
Query for all adverse events
}
