% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_clinical_observations}
\alias{all_clinical_observations}
\title{Query for all clinical observations}
\usage{
all_clinical_observations(
  dataset.name,
  clinical.observation.set.name = NULL,
  obs.type = NULL
)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{clinical.observation.set.name}{The name of the clinical-observation-set. All the observations in that
set will be pulled}

\item{obs.type}{A string. The type of the observation to return (e.g. \code{tumor-sum-diam}).
Only used if clinical.observation.set.name is \code{NULL}, in which case all the observation
of the given type will be returned.. Note that you should not include the
namespace (i.e. \code{:clinical-observation} in the string)}
}
\value{
Returns a query suitable for getting all clinical observations
  of a given type for a specified dataset
}
\description{
Query for all clinical observations
}
