% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{all_db_idents}
\alias{all_db_idents}
\title{Query to return all db-idents}
\usage{
all_db_idents()
}
\value{
Returns a query suitable for getting all the db-idents from the database
}
\description{
Query to return all db-idents
}
