% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{all_gdc_anatomic_sites}
\alias{all_gdc_anatomic_sites}
\title{Get all GDC anatomic sites}
\usage{
all_gdc_anatomic_sites()
}
\value{
Returns a query suitable for getting all GDC anatomic sites
}
\description{
Get all GDC anatomic sites
}
