% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{all_gene_symbols}
\alias{all_gene_symbols}
\title{Get all gene symbols from the database}
\usage{
all_gene_symbols()
}
\value{
Returns a query suitable for getting all gene symbols from the database
}
\description{
Get all gene symbols from the database
}
