% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{all_genes}
\alias{all_genes}
\title{Get all genes information from the database}
\usage{
all_genes()
}
\value{
Returns a query suitable for getting all the genes information
}
\description{
Get all genes information from the database
}
