% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{all_genomic_coordinates}
\alias{all_genomic_coordinates}
\title{Get all genomic-coordinate information from the database}
\usage{
all_genomic_coordinates()
}
\value{
Returns a query suitable for getting all the genomic coordinate
     information
}
\description{
Get all genomic-coordinate information from the database
}
