% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_meas_set_meas_types}
\alias{all_meas_set_meas_types}
\title{Query for all measurement types in a measurement set}
\usage{
all_meas_set_meas_types(dataset.name, measurement.set.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{measurement.set.name}{The name of the measurement set}
}
\value{
Returns a query suitable for getting all the above data for a given
  dataset
}
\description{
Query for all measurement types in a measurement set
}
