% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_meas_set_samples}
\alias{all_meas_set_samples}
\title{Query for all measurement sets and assays for all samples in a dataset}
\usage{
all_meas_set_samples(dataset.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}
}
\value{
Returns a query suitable for getting all the above data for a given
  dataset
}
\description{
Query for all measurement sets and assays for all samples in a dataset
}
