% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_meas_set_single_sample}
\alias{all_meas_set_single_sample}
\title{Query for all measurement sets and assays for one specific sample}
\usage{
all_meas_set_single_sample(dataset.name, sample.id)
}
\arguments{
\item{dataset.name}{The name of the dataset}
}
\value{
Returns a query suitable for getting all the above data for a given
  dataset and sample ID
}
\description{
Query for all measurement sets and assays for one specific sample
}
