% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_measurement_matrices}
\alias{all_measurement_matrices}
\title{Query to find all measurement matrices and backing files.}
\usage{
all_measurement_matrices()
}
\value{
Returns a query suitable to find all measurement matrix backing
files. Joins each matrix to its assay, dataset, measurement set context.
}
\description{
Query to find all measurement matrices and backing files.
}
