% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_measurements}
\alias{all_measurements}
\title{Query to return all measurements for a given measurement type}
\usage{
all_measurements(
  dataset.name,
  measurement.type,
  technology = NULL,
  assay.name = NULL,
  measurement.set.name = NULL,
  sel.epitopes = NULL,
  sel.cell.pops = NULL,
  include.targets = TRUE
)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{measurement.type}{The type of measurement, without the
\code{:measurement} namespace portion, ie \code{"nanostring-count"} or
\code{"percent-of-parent"}}

\item{technology}{DEPRECATED - this argument is unused and will be
removed in a future release.}

\item{assay.name}{Optional, if provided restricts the results to measurements
associated to the assay with the given name}

\item{measurement.set.name}{Optional, if provided restricts the results to
measurements associated with the measurement set with the given name}

\item{sel.epitopes}{Optional. Restrict the results to the specific epitope names}

\item{sel.cell.pops}{Optional. Restrict the results to the specified cell population names}

\item{include.targets}{Whether to include the targets of the measurement in
the result. The types of targets vary according to the measurement type,
for instance targets are genes for \code{measurement.type =
"nanostring-count"}, cell populations for \code{measurement.type =
"percent-of-leukocytes"} etc.}
}
\value{
Returns a query suitable to get all the measurements for a given
  measurement type
}
\description{
Query to return all measurements for a given measurement type
}
