% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_otus}
\alias{all_otus}
\title{Query to get all OTU's}
\usage{
all_otus(dataset.name, assay.name, measurement.set.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{assay.name}{The assay name}

\item{measurement.set.name}{The name of the measurement set}
}
\value{
Returns a query suitable for getting all OTU's for a specific measurement set
}
\description{
Query to get all OTU's
}
