% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_samples}
\alias{all_samples}
\title{Query for all samples}
\usage{
all_samples(dataset.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}
}
\value{
Returns a query suitable for getting all samples
  for a given dataset
}
\description{
Query for all samples
}
