% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_subjects}
\alias{all_subjects}
\title{Query for all subjects}
\usage{
all_subjects(dataset.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}
}
\value{
Returns a query suitable for getting all subjects
  for a given dataset
}
\description{
Query for all subjects
}
