% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_tcrs}
\alias{all_tcrs}
\title{Query to get all tcrs}
\usage{
all_tcrs(dataset.name, assay.name, measurement.set.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{assay.name}{The assay name}

\item{measurement.set.name}{The name of the measurement set}
}
\value{
Returns a query suitable for getting all the cell populations
}
\description{
Query to get all tcrs
}
