% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{all_timepoints}
\alias{all_timepoints}
\title{Query for all timepoints in a dataset}
\usage{
all_timepoints(dataset.name)
}
\arguments{
\item{dataset.name}{The name of the dataset}
}
\value{
Returns a query suitable for getting all timepoints in a given dataset
}
\description{
Query uses pull to include isolated timepoints with no regimen, and including
regimen name and relative order when applicable.
}
