% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{all_variants}
\alias{all_variants}
\title{Get all variant information from the database}
\usage{
all_variants()
}
\value{
Returns a query suitable for getting all the variant
     information
}
\description{
Get all variant information from the database
}
