% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{check_schema_compatibility}
\alias{check_schema_compatibility}
\title{Check schema compatibility}
\usage{
check_schema_compatibility(schema.ver)
}
\arguments{
\item{schema.ver}{The schema version (as returned by \link{get_db_schema_version})}
}
\description{
This function checks compatibility between the version of \code{wick} and the version
of the database schema, and issues a warning if they are not compatible
}
