% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{dataset_summary}
\alias{dataset_summary}
\title{Query to get a summary of assays, assay description, and measurement sets in a dataset}
\usage{
dataset_summary(dataset.name, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
A query suitable for returning a summary of dataset assays
}
\description{
Query to get a summary of assays, assay description, and measurement sets in a dataset
}
