% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{deduplicate_otu_taxonomic_levels}
\alias{deduplicate_otu_taxonomic_levels}
\title{Deduplicate otu taxonomic levels}
\usage{
deduplicate_otu_taxonomic_levels(all.otus, na.value, sep = "_")
}
\arguments{
\item{all.otus}{A table with OTU information as returned by \link{get_all_otus}}

\item{na.value}{Name to use for OTU's that have an \code{NA} at a specific taxonomic level}

\item{sep}{Character separator used to concatenate names if necessary}
}
\value{
Returns the data in \code{all.otus} with the names at each taxonomic level processed
  as necessary
}
\description{
This function processes otu information to make sure that names at each taxonomic
level are unique. If that's the case, names are left unchanged, otherwise
the name is transformed to be a concatenation of the names of upstream taxonomic levels,
to guarantee uniqueness. This function is used internally by \code{aggregate_otu_measurements},
and it is unlikely that you would have to call it directly as an end-user
}
