% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{do_query}
\alias{do_query}
\title{Runs a query against the database}
\usage{
do_query(query, db.name = NULL, ...)
}
\arguments{
\item{db.name}{Optional. The name of the database against which the
query is run. This can be set globally using \code{\link{set_dbname}}}

\item{...}{Additional arguments. Please refer to the documentation of
\code{datalogr::do_query}}
}
\value{
Returns the query results. Refer to the documentation of
  \code{datalogr::do_query} for more details
}
\description{
Runs a query against the database
}
