% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{ensure_dbname}
\alias{ensure_dbname}
\title{Ensure db.name has been set correctly}
\usage{
ensure_dbname(db.name)
}
\arguments{
\item{db.name}{The name of the database, possibly \code{NULL}}
}
\value{
Returns \code{db.name} if not \code{NULL}. Otherwise returns the global
  \code{pkg.env$db.name} value
}
\description{
Ensure db.name has been set correctly
}
