% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{ensure_molten}
\alias{ensure_molten}
\title{Ensure that data is molten}
\usage{
ensure_molten(meas.table, variable.name = "variable")
}
\arguments{
\item{meas.table}{Input \code{data.frame} as returned for instance from
\code{\link{get_all_measurements}}}

\item{variable.name}{Only used if the data is not molten already. This will be passed
to \code{reshape2::melt} and represents the name of variable used to store measured variable names}
}
\value{
Returns the input data in molten format (or unchanged if it was molten already)
}
\description{
This function ensures that data returned from e.g. \code{\link{get_all_measurements}}
is in molten format
}
