% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{flatten_subject_therapies}
\alias{flatten_subject_therapies}
\title{Flatten therapies information}
\usage{
flatten_subject_therapies(tab)
}
\arguments{
\item{tab}{Input \code{data.frame}. It must contain a column called \code{subject.therapies}}
}
\value{
Returns the input data with the \code{subject.therapies} column simplified to
  a string concatenated with \code{;}
}
\description{
This function flattens and simplifies therapy information for an input table. This information
is typically in a nested \code{data.frame} format. This function concatenates that information
into a plain string
}
