% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_adverse_events}
\alias{get_all_adverse_events}
\title{Return all adverse events for a given dataset}
\usage{
get_all_adverse_events(dataset.name, clinical.observation.set.name = NULL, ...)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{clinical.observation.set.name}{The name of the clinical-observation-set. All the observations in that
set will be pulled}
}
\value{
Returns a \code{data.frame} containing all the clinical observation
  and their associated subjects and timepoints for a given dataset
}
\description{
Return all adverse events for a given dataset
}
