% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_cell_populations}
\alias{get_all_cell_populations}
\title{Get all cell populations}
\usage{
get_all_cell_populations(dataset.name, assay.name, measurement.set.name, ...)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{assay.name}{The assay name}

\item{measurement.set.name}{The name of the measurement set}

\item{...}{Additional parameters passed to \code{do_query}}
}
\value{
Returns a character vector with all the cell populations names
}
\description{
Get all cell populations
}
