% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_cell_types}
\alias{get_all_cell_types}
\title{Get all cell types from the database}
\usage{
get_all_cell_types(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a character vector with all the cell types in the
  database
}
\description{
Get all cell types from the database
}
