% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_clinical_observations}
\alias{get_all_clinical_observations}
\title{Return all clinical observations for a given dataset}
\usage{
get_all_clinical_observations(
  dataset.name,
  clinical.observation.set.name = NULL,
  obs.type = NULL,
  ...
)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{clinical.observation.set.name}{The name of the clinical-observation-set. All the observations in that
set will be pulled}

\item{obs.type}{A string. The type of the observation to return (e.g. \code{tumor-sum-diam}).
Only used if clinical.observation.set.name is \code{NULL}, in which case all the observation
of the given type will be returned.. Note that you should not include the
namespace (i.e. \code{:clinical-observation} in the string)}
}
\value{
Returns a \code{data.frame} containing all the clinical observation
  and their associated subjects and timepoints for a given dataset
}
\description{
Return all clinical observations for a given dataset
}
