% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_cnv}
\alias{get_all_cnv}
\title{Get all cnv from the database}
\usage{
get_all_cnv(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a data.frame with all cnv information
}
\description{
Get all cnv from the database
}
