% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_db_idents}
\alias{get_all_db_idents}
\title{Get all db-idents}
\usage{
get_all_db_idents(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a \code{data.frame} with all the db-idents in the database
  and their associated entity id
}
\description{
Get all db-idents
}
