% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_epitopes}
\alias{get_all_epitopes}
\title{Get all epitopes from the database}
\usage{
get_all_epitopes(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a character vector with all the epitope ids in the database
}
\description{
Get all epitopes from the database
}
