% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_gdc_anatomic_sites}
\alias{get_all_gdc_anatomic_sites}
\title{Get all GDC anatomic sites from the database}
\usage{
get_all_gdc_anatomic_sites(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a character vector containing all the GDC anatomic sites in the database
}
\description{
Get all GDC anatomic sites from the database
}
