% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_gene_products}
\alias{get_all_gene_products}
\title{Get all gene product information from the database}
\usage{
get_all_gene_products(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a \code{data.frame} with complete information about all the gene
  entities in the database (including alternate names and symbols)
}
\description{
Get all gene product information from the database
}
