% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_gene_symbols}
\alias{get_all_gene_symbols}
\title{Get all gene symbols from the database}
\usage{
get_all_gene_symbols(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a character vector with all the HGNC gene symbols in the database
}
\description{
Get all gene symbols from the database
}
