% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_meas_set_meas_types}
\alias{get_all_meas_set_meas_types}
\title{For a given measurement-set, return all measurement types
of measurements in that set}
\usage{
get_all_meas_set_meas_types(
  dataset.name,
  measurement.set.name,
  simple.strings = TRUE,
  ...
)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{measurement.set.name}{The name of the measurement set}

\item{simple.strings}{reformats the datomic entity types to strings
containing just the name of the measurement type.}
}
\value{
Returns a \code{vector} containing all the measurement types
of measurements in the given measurement-set.
}
\description{
For a given measurement-set, return all measurement types
of measurements in that set
}
