% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_meas_set_measurement_attributes}
\alias{get_all_meas_set_measurement_attributes}
\title{For a given measurement-set, return all attributes
of measurements in that set}
\usage{
get_all_meas_set_measurement_attributes(
  dataset.name,
  measurement.set.name,
  ...
)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{measurement.set.name}{The name of the measurement set}
}
\value{
Returns a \code{vector} containing all the attributes
of measurements in the given measurement-set.
}
\description{
For a given measurement-set, return all attributes
of measurements in that set
}
