% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_meas_set_samples}
\alias{get_all_meas_set_samples}
\title{For a given datatset, return all measurement sets and the samples and
timepoints they measure for a given dataset}
\usage{
get_all_meas_set_samples(dataset.name, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a \code{data.frame} containing information about all the
  measurement sets in the dataset, their asssays, and the samples they
  measure, joined with information about the subjects they originate from and
  treatment regimens associated with those subjects
}
\description{
For a given datatset, return all measurement sets and the samples and
timepoints they measure for a given dataset
}
