% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_measurement_matrices}
\alias{get_all_measurement_matrices}
\title{Retrieve a table of measurement matrices in a database, with information on
dataset, assay, and measurement-set context for retrieving the full matrix
with \code{get_measurement_matrix}.}
\usage{
get_all_measurement_matrices(...)
}
\arguments{
\item{...}{Additional parameters passed to \code{do_query}}
}
\value{
All measurement matrices in a database.
}
\description{
Retrieve a table of measurement matrices in a database, with information on
dataset, assay, and measurement-set context for retrieving the full matrix
with \code{get_measurement_matrix}.
}
