% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_measurements}
\alias{get_all_measurements}
\title{Get all measurements for a given measurement type}
\usage{
get_all_measurements(
  dataset.name,
  measurement.type,
  assay.name = NULL,
  technology = NULL,
  measurement.set.name = NULL,
  include.targets = TRUE,
  cast.output = TRUE,
  fun.aggregate = NULL,
  include.sample.context = FALSE,
  group.by.assay.meas.set = TRUE,
  sel.epitopes = NULL,
  sel.cell.pops = NULL,
  ...
)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{measurement.type}{The type of measurement, without the
\code{:measurement} namespace portion, ie \code{"nanostring-count"} or
\code{"percent-of-parent"}}

\item{assay.name}{Optional, if provided restricts the results to measurements
associated to the assay with the given name}

\item{technology}{DEPRECATED - this argument is unused and will be
removed in a future release.}

\item{measurement.set.name}{Optional, if provided restricts the results to
measurements associated with the measurement set with the given name}

\item{include.targets}{Whether to include the targets of the measurement in
the result. The types of targets vary according to the measurement type,
for instance targets are genes for \code{measurement.type =
"nanostring-count"}, cell populations for \code{measurement.type =
"percent-of-leukocytes"} etc.}

\item{cast.output}{Whether the resulting data should be \code{cast} using \code{reshape2}
or kept in \code{melt} format.}

\item{fun.aggregate}{Only used if \code{cast.output == TRUE}. Function to aggregate
across multiple measurements with the same sample and target.
Defaults to \code{NULL} and will emit a warning and
use \code{length} if an aggregation function is needed but not supplied.}

\item{include.sample.context}{Whether to include the sample context in the results
using \code{link{get_sample_context}}. Note that if this option is specified
the output will be molten, irrespective of the value of \code{cast.output}}

\item{sel.epitopes}{Optional. Restrict the results to the specific epitope names}

\item{sel.cell.pops}{Optional. Restrict the results to the specified cell population names}

\item{...}{Additional parameters passed to \code{do_query}}
}
\value{
Returns a named list of datasets, one for each assay and
  measurement set combination, or a single dataset if there
  is only one such combinations. Each dataset is either a \code{matrix},
  if \code{cast.output == FALSE} or a \code{data.frame} otherwise
}
\description{
Get all measurements for a given measurement type
}
