% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_otus}
\alias{get_all_otus}
\title{Get all OTU's}
\usage{
get_all_otus(dataset.name, assay.name, measurement.set.name, ...)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{assay.name}{The assay name}

\item{measurement.set.name}{The name of the measurement set}

\item{...}{Additional parameters passed to \code{do_query}}
}
\description{
Get all OTU's for a specific measurement set
}
