% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_proteins}
\alias{get_all_proteins}
\title{Get all proteins from the database}
\usage{
get_all_proteins(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a character vector with all the protein names in the database
}
\description{
Get all proteins from the database
}
