% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_samples}
\alias{get_all_samples}
\title{Return all samples information for a given dataset}
\usage{
get_all_samples(dataset.name, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a \code{data.frame} containing information about all the samples
  in the dataset, joined with information about the subjects they originate
  from and treatment regimens associated with those subjects
}
\description{
Return all samples information for a given dataset
}
