% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_subjects}
\alias{get_all_subjects}
\title{Return all subject information for a given dataset}
\usage{
get_all_subjects(dataset.name, include.response.data = TRUE, ...)
}
\arguments{
\item{include.response.data}{Boolean. If TRUE (default), also pulls response
data for all subjects, including PFS, OS, and Best Overall Response (bor).
Will error if it finds more than one PFS or OS clinical-observation entity
per subject. Current calculates BOR rather than pulling it from the
database.}

\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\description{
This function gets information on all subjects in a dataset, including
demographics, treatment regimen, and (optionally) response information. If
response data is included, this function will error if multiple entities
exist for clinical-observations where we expect a single value, ie multiple
pfs or os values.
}
