% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_all_tcrs}
\alias{get_all_tcrs}
\title{Get all TCR data.}
\usage{
get_all_tcrs(dataset.name, assay.name, measurement.set.name, ...)
}
\arguments{
\item{dataset.name}{The name of the dataset}

\item{assay.name}{The assay name}

\item{measurement.set.name}{The name of the measurement set}

\item{...}{Additional parameters passed to \code{do_query}}
}
\value{
Returns a character vector with all the tcr names
}
\description{
Get all TCR data.
}
