% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_reference_data.R
\name{get_all_variants}
\alias{get_all_variants}
\title{Get all variant information from the database}
\usage{
get_all_variants(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a \code{data.frame} with complete information about all the
  variant entities in the database
}
\description{
Get all variant information from the database
}
