% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{get_candelabra_oauth_info}
\alias{get_candelabra_oauth_info}
\title{Get the oauth url from candelabra}
\usage{
get_candelabra_oauth_info()
}
\value{
Returns a list with two elements
  \itemize{
    \item oauth.url The oauth url
    \item redirect.url The redirect url
  }
}
\description{
Get the oauth url from candelabra
}
