% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_common_timepoints}
\alias{get_common_timepoints}
\title{Return timepoint information}
\usage{
get_common_timepoints(dataset.name, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{\link{do_query}} (typically
\code{db.name}, unless it has been set globally)}
}
\value{
Returns a \code{data.frame} with timepoint information
}
\description{
This function returns timepoint information for timepoints that are common to all
the treatment regimen. The timepoints are renamed by removing
the portion before the \code{/} separator. This function
is intended to generate simple timepoint information which is valid across
all treatment regimens for cases where all the treatment regimens have the
same timepoint structure. If the timepoint structure is not identical across all
treatment.regimens, this function will issue a warning. Note that in this case
the relative order of timepoints will be inconsistent across different treatment regimens.
THIS FUNCTIONALITY WILL BE REMOVED IN THE FUTURE AND ONLY COMMON TIMEPOINTS WILL BE RETURNED
}
