% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{get_db_schema_version_via_query}
\alias{get_db_schema_version_via_query}
\title{Return the database schema version via a query}
\usage{
get_db_schema_version_via_query(db.name)
}
\arguments{
\item{db.name}{The name of the database}
}
\value{
Returns the schema version as a string
}
\description{
Return the database schema version via a query
}
