% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{get_matrix_file}
\alias{get_matrix_file}
\title{Return a matrix file as a data.frame.}
\usage{
get_matrix_file(matrix_file, db.name = NULL)
}
\arguments{
\item{db.name}{Optional: the name of the database}

\item{matrix.file}{The matrix file to retrieve.}
}
\value{
Returns the schema version as a string
}
\description{
Return a matrix file as a data.frame.
}
