% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_measurement_matrices}
\alias{get_measurement_matrices}
\title{Get all the measurement matrix metadata for a given dataset, assay,
measurement set, and measurement matrix (name) context.}
\usage{
get_measurement_matrices(dataset.name, assay.name, measurement.set.name, ...)
}
\arguments{
\item{dataset.name}{The dataset that this measurement matrix is part of.}

\item{assay.name}{The assay this measurement matrix corresponds to.}

\item{measurement.set.name}{The measurement set that includes this
measurement matrix.}

\item{...}{Additional parameters passed to \code{do_query}}
}
\value{
All measurement matrices in a database.
}
\description{
Get all the measurement matrix metadata for a given dataset, assay,
measurement set, and measurement matrix (name) context.
}
