% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_measurement_matrix}
\alias{get_measurement_matrix}
\title{Get the full measurement matrix for a given dataset, assay, measurement set,
and measurement matrix (name) context.}
\usage{
get_measurement_matrix(
  dataset.name,
  assay.name,
  measurement.set.name,
  measurement.matrix.name,
  ...
)
}
\arguments{
\item{dataset.name}{The dataset that this measurement matrix is part of.}

\item{assay.name}{The assay this measurement matrix corresponds to.}

\item{measurement.set.name}{The measurement set that includes this
measurement matrix.}

\item{measurement.matrix.name}{The name of the measurement matrix to find
the backing file for.}

\item{...}{Additional parameters passed to \code{do_query}}
}
\value{
All measurement matrices in a database.
}
\description{
Get the full measurement matrix for a given dataset, assay, measurement set,
and measurement matrix (name) context.
}
