% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{get_schema_compatibility_overrides}
\alias{get_schema_compatibility_overrides}
\title{#' @return Returns the named list of schema versions that are compatible
          with each other.}
\usage{
get_schema_compatibility_overrides()
}
\description{
#' @return Returns the named list of schema versions that are compatible
          with each other.
}
