% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{get_unified_timeline}
\alias{get_unified_timeline}
\title{Return a unified timeline for mulitple treatment regimens}
\usage{
get_unified_timeline(dataset.name, treatment.regimen.name = NULL, ...)
}
\arguments{
\item{treatment.regimen.name}{optional character vector. If provided, only limit
the timeline to the selected treatment regimens}

\item{...}{Additional arguments passed to \code{get_all_timepoints}}
}
\value{
Returns a \code{data.frame} with timepoint information
}
\description{
This function attempts to construct a unified timeline for multiple treatment regimens.
It assumes a situation where timepoints that are conceptually similar between multiple
treatment regimens have the same name, but potentially each treatment regimen has additional
timepoints interspersed that are not shared with other regimens. The unified timeline
is constructed by taking the union of all the timepoint ids, and, for each timepoint id,
selecting the maximum among all its relative orders across the treatment regimens
}
