% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{group_by_assay_meas_set}
\alias{group_by_assay_meas_set}
\title{Group data by assay and measurement set}
\usage{
group_by_assay_meas_set(
  tab,
  measure.var,
  target.var = NULL,
  cast.output = TRUE,
  fun.aggregate = NULL
)
}
\arguments{
\item{tab}{The query result}

\item{measure.var}{The name of the column in \code{tab} representing
the measurement value (e.g. \code{"count"})}

\item{target.var}{Optional. The name(s) of the column(s) in \code{tab} representing
the target of the measurement (e.g. \code{"gene.hugo"} or \code{c("cell.pop", "epitope.id")})
If not specified, the result will only contain the data in \code{measure.var}. Must be supplied
if \code{cast.output} is specified}

\item{cast.output}{Whether the resulting data should be \code{cast} using \code{reshape2}
or kept in \code{melt} format.}

\item{fun.aggregate}{Only used if \code{cast.output == TRUE}. Function to aggregate
across multiple measurements with the same sample and target.
Defaults to \code{NULL} and will emit a warning and
use \code{length} if an aggregation function is needed but not supplied.}
}
\value{
Returns a named list of datasets, one for each assay and
  measurement set combination, or a single dataset if there
  is only one such combinations. Each dataset is either a \code{matrix},
  if \code{cast.output == FALSE} or a \code{data.frame} otherwise
}
\description{
This function groups a measurements query result by
assay and measurement set
}
