% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{login}
\alias{login}
\title{Login into the CANDEL system}
\usage{
login(email = NULL, force = FALSE)
}
\arguments{
\item{email}{DEPRECATED. The user email}

\item{force}{By default this function reads a token cached in the \code{.candel}
directory in the user home. If this token has expired, you can pass the
\code{force} option to force a refresh}
}
\description{
This function is used to login into the CANDEL system. It only needs
to be invoked once per session (unless the token expires, in which
case it needs to be called again, see below). Upon calling this function,
if a cached authorization token is not already available,
a browser window will open asking the user to authenticate into his/her
account. When this is done, a page will display an authorization code
that needs to be entered when prompted by the login function, thus
completing the login process
}
