% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{map_gene_symbols}
\alias{map_gene_symbols}
\title{Maps gene symbols}
\usage{
map_gene_symbols(all.genes, gene.symbol, suppress.warnings = TRUE)
}
\arguments{
\item{all.genes}{Information about all the genes as returned by \link{get_all_genes}}

\item{gene.symbol}{Character vector of symbols to map}

\item{suppress.warnings}{Boolean suppressing warning messages indicating the number of successfully mapped genes}
}
\value{
Returns a named vector where the names are the original symbols and the corresponding values
  the ones they have been mapped to
}
\description{
This function maps gene symbols on the HGNC symbols used within the database.
Note that if you have multiple symbols to match, it is quicker to call this function once
with all the symbols, as opposed to calling it multiple times with the individual symbols
}
