% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{matrix_backing_file}
\alias{matrix_backing_file}
\title{Query to find all measurement matrices and backing files.}
\usage{
matrix_backing_file(
  dataset.name,
  assay.name,
  measurement.set.name,
  measurement.matrix.name,
  ...
)
}
\arguments{
\item{dataset.name}{The dataset that this measurement matrix is part of.}

\item{assay.name}{The assay this measurement matrix corresponds to.}

\item{measurement.set.name}{The measurement set that includes this
measurement matrix.}

\item{measurement.matrix.name}{The name of the measurement matrix to find
the backing file for.}
}
\value{
Returns a query suitable to find all measurement matrix backing
files. Joins each matrix to its assay, dataset, measurement set context.
}
\description{
Query to find all measurement matrices and backing files.
}
