% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries_dataset.R
\name{measurement_matrices}
\alias{measurement_matrices}
\title{Query to find a measurement matrix backing file. The backing file is
the key that needs to be passed to candelabra to retrieve a full
measurement matrix file.}
\usage{
measurement_matrices(dataset.name, assay.name, measurement.set.name)
}
\arguments{
\item{dataset.name}{The dataset that this measurement matrix is part of.}

\item{assay.name}{The assay this measurement matrix corresponds to.}

\item{measurement.set.name}{The measurement set that includes this
measurement matrix.}
}
\value{
Returns a query suitable to find a measurement matrix backing file.
}
\description{
Query to find a measurement matrix backing file. The backing file is
the key that needs to be passed to candelabra to retrieve a full
measurement matrix file.
}
