% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{resolve_db_idents}
\alias{resolve_db_idents}
\title{Resolve db-idents in input data}
\usage{
resolve_db_idents(tab, col.names, ...)
}
\arguments{
\item{tab}{The input data}

\item{col.names}{A character of vector of column names in \code{tab} that must be processed}
}
\value{
Returns \code{tab} with the specified columns processed
}
\description{
This function resolves columns containing db id's pointing to db-idents,
by substituting the values with the names of the idents that the id's point to.
This is useful for instance if you have a column of numeric id's in your data that represent
the id's of database enums, and you want to substitute them with the actual enums themselves
}
