% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{select_assay_exp}
\alias{select_assay_exp}
\title{Select single assay and measurement set from a list of measurement query results}
\usage{
select_assay_exp(tab, assay.name = NULL, measurement.set.name = NULL)
}
\arguments{
\item{tab}{A \code{data.frame} returned from a "get_all_X_measurements" query function}

\item{assay.name}{The name of the assay. Required if there are multiple assays for this data type in this dataset}

\item{measurement.set.name}{The name of the measurement set. Required if there are multiple measurement sets
and / or multiple assays in this dataset}
}
\value{
Returns a \code{data.frame} of data corresponding to a specific
  assay and measurement set combination, or \code{tab} if they are
  both \code{NULL}
}
\description{
Select single assay and measurement set from a list of measurement query results
}
