% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{set_dbname}
\alias{set_dbname}
\title{Globally set the database name}
\usage{
set_dbname(db.name)
}
\arguments{
\item{db.name}{If using the pret provision workflow, this is the database
name, ie "laceys-database". If using the pret request-db workflow, this is
the query url sent to you in your email notification, ie
"http://ec2-54-80-237-178.compute-1.amazonaws.com/query/cdel"}
}
\value{
Returns the previous \code{db.name} value, invisibly
}
\description{
This function sets a database name globally for the entire package (so that you
  don't have to pass around a \code{db.name} argument everywhere)
}
