% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{set_query_server}
\alias{set_query_server}
\title{Sets the query server}
\usage{
set_query_server(query.server)
}
\arguments{
\item{query.server}{The query server address}
}
\value{
Returns the previous \code{query.server} value, invisibly
}
\description{
This function is for development purposes only. Sets the query server for this
session. The query server for the old Dyanmo stack is
\code{http://parkerici-cdel-query.us-east-1.elasticbeanstalk.com}
}
