% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wick.R
\name{set_token}
\alias{set_token}
\title{Globally set the auth token}
\usage{
set_token(token)
}
\arguments{
\item{token}{Character. The token}
}
\value{
Returns the previous token, invisibly
}
\description{
This function globally sets the authorization token. Users should use
\code{\link{login}}, instead of calling this directly
}
